[NWPS-1376] - Moved component into correct directory structure
