+++ app/assets/hee/blocks/content/footer/anchorlinks-sticky/_anchorlinks-sticky.scss
