+++ app/assets/hee/blocks/content/footer/anchorlinks-sticky/macro.njk
@@ -1,4 +1,4 @@
-{%- from 'components/hee/hee-anchorlinks/macro.njk' import anchorLinks %}
+{%- from 'hee/blocks/content/sidebar/anchorlinks/macro.njk' import anchorLinks %}
 
 {% macro anchorLinksSticky(params) %}
   <div aria-expanded="false" class="hee-anchorlinks__sticky" id="hee-anchorlinks-sticky">
