+++ app/assets/hee/hee.scss
@@ -61,7 +61,7 @@
 @import '../components/hee-table-2';
 @import 'blocks/content/main/applies-to/applies-to';
 @import 'blocks/content/sidebar/anchorlinks/anchorlinks';
-@import '../components/hee/hee-anchorlinks-sticky/anchorlinks-sticky';
+@import 'blocks/content/footer/anchorlinks-sticky/anchorlinks-sticky';
 @import 'blocks/scaffolding/notifications/phasebanner/phasebanner';
 @import '../components/hee/hee-tabs/tabs';
 @import '../components/nhsuk-hee/nhsuk-hee-tables/tables2'; //?
